Split ipynb functions
