+++ lib/dataprocessor/pipes/ipynb.py
@@ -1,34 +1,8 @@
 # -*- coding: utf-8 -*-
 
 
-import psutil
 import webbrowser
-from ..utility import check_file
-from ..exception import DataProcessorError as dpError
-
-
-def gather_notebooks():
-    """ Gather processes of IPython Notebook
-    """
-    notes = []
-    for p in psutil.process_iter():
-        if not p.name().lower() in ["ipython", "python"]:
-            continue
-        if "notebook" not in p.cmdline():
-            continue
-        for net in p.connections(kind="inet4"):
-            if net.status != "LISTEN":
-                continue
-            _, port = net.laddr
-            break
-        notes.append({
-            "pid": p.pid,
-            "cwd": p.cwd(),
-            "port": port,
-        })
-    if not notes:
-        raise dpError("No IPython Notebook found")
-    return notes
+from ..ipynb import resolve_url
 
 
 def start(nl, ipynb_path):
@@ -50,20 +24,12 @@ def start(nl, ipynb_path):
         - Cannot open browser
 
     """
-    ipynb_path = check_file(ipynb_path)
-    for note in gather_notebooks():
-        cwd = note["cwd"]
-        if not ipynb_path.startswith(cwd):
-            continue
-        note["postfix"] = ipynb_path[len(cwd) + 1:]  # remove '/'
-        url = "http://localhost:{port}/notebooks/{postfix}".format(**note)
-        try:
-            webbrowser.open(url)
-        except webbrowser.Error:
-            raise dpError("Unknown error while opening notebook")
-        return nl
-    raise dpError("No valid Notebook found. "
-                  "Please stand notebook server on the parent directory.")
+    url = resolve_url(ipynb_path)
+    try:
+        webbrowser.open(url)
+    except webbrowser.Error:
+        raise dpError("Unknown error while opening notebook")
+    return nl
 
 
 def register(pipes_dics):
